feat: extract and export `estimate_tx_compressed_size`
